requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
twilio==8.13.0
groq==0.12.0
httpx==0.24.1
PyMySQL==1.1.0
deep-translator==1.11.4
python-dateutil==2.8.2
requests==2.31.0